Revert "Reduce number of headers exposed for JSStringDecoder"
This reverts commit e55a53ee0ff0379acbd76b1f1722aa40c54afcfc.
